Fix the bug of a logic problem
In  previous DbscanModel.scala file, while building the new graph with core points and edges, the previous version will build a graph with noise part. It can only filter those noise points with outDegree==0, but it cannot filter those parts, for instance, including only 2 close points. The number of the points of this ‘cluster’ is 2 which may be smaller than the minNumOfPoints while they will be still counted as one final cluster.